Use when_first_matching_example_defined
require less files depending on spec type
